Refactor QR code data segment classes to improve encoding logic and structure
